test(http2): move options-getter rehash repro to a fixture file
The "does not hold *Stream across user-controlled options getters" test
spawned its subprocess via `bun -e`, which on a debug+ASAN build runs the
129 concurrent h2 requests in ~6.3s and overruns the default 5s per-test
timeout. The other two tests in the same file spawn fixture files and
finish in ~3s.

Move the script to node-http2-getter-rehash.fixture.js to match, and cut
the inner-loop stream count from 128 to 32 (the streams HashMap starts at
capacity 8 and grows at 80% load, so 32 still forces several rehashes).
Drop parent/weight from the options: they now emit DEP0194 and are
deleted before the native call anyway.

Before: 3/3 timeouts at 5s. After: 3/3 pass in ~3.2s.
